requirements.txt
openai
python-dotenv
fastapi[standard]
python-multipart